requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
cryptography==37.0.4
distlib==0.3.5
dj-database-url==0.5.0
Django==4.0.6
django-rest-knox==4.2.0
djangorestframework==3.13.1
filelock==3.7.1
pipenv==2022.7.4
platformdirs==2.5.2
psycopg2==2.9.3
pycparser==2.21
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
virtualenv==20.15.1
virtualenv-clone==0.5.7
